requirements.txt
asposestorage==1.0.2
matplotlib==3.5.1
numpy==1.25.2
numpy==1.21.5
opencv_python==4.7.0.72
Pillow==9.0.1
Pillow==10.0.0
pytorch_lightning==2.0.5
torch==2.0.1+cu118
torchmetrics==0.6.0
torchvision==0.15.2+cu118
